lblj/baseline/requirements.txt
python>=3.7
jieba==0.42.1
pytorch==1.12.0
transformers==4.16.2
